% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA_test.R
\name{anova_test}
\alias{anova_test}
\title{Performs genewise ANOVA test on a cell x row matrix}
\usage{
anova_test(data)
}
\arguments{
\item{data}{matrix where cells are rows and genes are columns}
}
\value{

}
\description{
Performs genewise ANOVA test on a cell x row matrix
}
