% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_test.R
\name{batchLRT_linearModel}
\alias{batchLRT_linearModel}
\title{Linear Model LRT for dose dependent differential gene expression}
\usage{
batchLRT_linearModel(sce)
}
\arguments{
\item{sce}{A single cell object (Does not allow a logcount matrix with genes having
a dropout of 100\%. Such genes should be filtered out before running the test)}
}
\value{
A data frame giving the test statistic and unadjusted p values
for the linear, logistic and combined regression models for p genes
}
\description{
Linear Model LRT for dose dependent differential gene expression
}
\author{
Satabdi Saha
}
