% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes_factor.R
\name{Bayes_factor}
\alias{Bayes_factor}
\title{Calculation of Bayes factor}
\usage{
Bayes_factor(
  Y_t,
  Y_c,
  m_t0,
  m_c0,
  m_0,
  tau_t_mu,
  tau_c_mu,
  tau_mu,
  b_sigma,
  a_sigma,
  a_w,
  a_c_w,
  a_t_w,
  b_w,
  b_t_w,
  b_c_w,
  prior_alt,
  prior_null
)
}
\arguments{
\item{Y_t}{list of matrices of observed data having n cells and p genes}

\item{Y_c}{list of matrices of observed data having n cells and p genes
in the control group}

\item{m_t0}{INSERT DESCR}

\item{m_c0}{INSERT DESCR}

\item{m_0}{INSERT DESCR}

\item{tau_t_mu}{INSERT DESCR}

\item{tau_c_mu}{INSERT DESCR}

\item{tau_mu}{INSERT DESCR}

\item{b_sigma}{INSERT DESCR}

\item{a_sigma}{INSERT DESCR}

\item{a_w}{INSERT DESCR}

\item{a_c_w}{INSERT DESCR}

\item{a_t_w}{INSERT DESCR}

\item{b_w}{INSERT DESCR}

\item{b_t_w}{INSERT DESCR}

\item{b_c_w}{INSERT DESCR}

\item{prior_alt}{INSERT DESCR}

\item{prior_null}{INSERT DESCR}
}
\value{
INSERT DESCR
}
\description{
Calculation of Bayes factor
}
\details{

}
