% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes_test.R
\name{Bayes_factor_multiple}
\alias{Bayes_factor_multiple}
\title{Title Bayesian Test for dose-dependent differential gene expression analysis}
\usage{
Bayes_factor_multiple(Y, prior, detailed = FALSE)
}
\arguments{
\item{Y}{list of size K (dose levels) of vectors length n (number of cells).}

\item{prior}{priors object generated with sceCalcPriors.}

\item{detailed}{logical output detailed statistical estimates.}
}
\value{
output A list having the estimates of the log-likelihood function,
log prior odds, and the log Bayes factor test statistic.
}
\description{
Title Bayesian Test for dose-dependent differential gene expression analysis
}
\author{
Satabdi Saha
}
