% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes_test.R, R/retired_functions.R
\name{bayesDETest}
\alias{bayesDETest}
\title{INSERT DESCRIPTION}
\usage{
bayesDETest(data.list, m, tau_k_mu, tau_mu, prior_Alter, prior_Null)

bayesDETest(data.list, m, tau_k_mu, tau_mu, prior_Alter, prior_Null)
}
\arguments{
\item{priors}{INSERT DESCRIPTION}

\item{detailed}{A boolean denoting how much info to show in result}

\item{sce}{SingleCellExperiment object with a logcounts assay
and Dose column in the cell metadata}
}
\value{
INSESRT DESCRIPTION - describe omega etc...

a vector of p values from the ANOVA test
}
\description{
INSERT DESCRIPTION

Performs a genewise ANOVA test on a SingleCellExperiment object
}
