% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retired_functions.R
\name{bf_pair}
\alias{bf_pair}
\title{Performs a genewise ANOVA test on a SingleCellExperiment object}
\usage{
bf_pair(data.list, bf_1, m, tau_k_mu, tau_mu, prior_Alter, prior_Null)
}
\arguments{
\item{sce}{SingleCellExperiment object with a logcounts assay
and Dose column in the cell metadata}
}
\value{
a vector of p values from the ANOVA test
}
\description{
Performs a genewise ANOVA test on a SingleCellExperiment object
}
