% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newBayesTest.R
\name{calc_a_sigma_b_sigma}
\alias{calc_a_sigma_b_sigma}
\title{Calculate a_sigma, b_sigma}
\usage{
calc_a_sigma_b_sigma(sce)
}
\arguments{
\item{sce}{Single Cell Object}
}
\value{
a_sigma and b_sigma values
}
\description{
Calculate a_sigma, b_sigma
}
\author{
Satabdi Saha
}
