% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prior_hyperparameters.R
\name{calc_a_w_b_w}
\alias{calc_a_w_b_w}
\title{Calculate a_w, b_w}
\usage{
calc_a_w_b_w(omega_mean, omega_var)
}
\arguments{
\item{omega_mean}{INSERT DESCR HERE}

\item{omega_var}{INSERT DESCR HERE}
}
\value{
a_w and b_w values
}
\description{
Calculate a_w, b_w
}
\author{
Satabdi Saha
}
