% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes_test.R
\name{calc_alt_null}
\alias{calc_alt_null}
\title{Calculate a_w, b_w}
\usage{
calc_alt_null(sce, method = "fixed", de.prob = 0.25)
}
\arguments{
\item{omega_mean}{Groupwise dropout means}

\item{omega_var}{Groupwise dropout variance}
}
\value{
a_w and b_w values
}
\description{
Calculate a_w, b_w
}
\author{
Satabdi Saha
}
