% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retired_functions.R
\name{calc_priors}
\alias{calc_priors}
\title{Performs a genewise ANOVA test on a SingleCellExperiment object}
\usage{
calc_priors(simulated.data, simulated.cell.metadata, simulated.gene.metadata)
}
\arguments{
\item{sce}{SingleCellExperiment object with a logcounts assay
and Dose column in the cell metadata}
}
\value{
a vector of p values from the ANOVA test
}
\description{
Performs a genewise ANOVA test on a SingleCellExperiment object
}
