% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retired_functions.R
\name{calcFC}
\alias{calcFC}
\title{Performs a genewise ANOVA test on a SingleCellExperiment object}
\usage{
calcFC(sim)
}
\arguments{
\item{sce}{SingleCellExperiment object with a logcounts assay
and Dose column in the cell metadata}
}
\value{
a vector of p values from the ANOVA test
}
\description{
Performs a genewise ANOVA test on a SingleCellExperiment object
}
