% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes_test.R
\name{calculate_BF}
\alias{calculate_BF}
\title{Title Compute Bayes factors for all the p-genes}
\usage{
calculate_BF(Y, par_null, par_alt, prior.null, prior.alt)
}
\arguments{
\item{par_null}{: Parameters obtained after optimizing the marginal likelihood under the null hypothesis}

\item{par_alt}{: Parameters obtained after optimizing the marginal likelihood under the alternative hypothesis}

\item{prior.null}{: Prior probability of the null hypothesis}

\item{prior.alt}{: Prior probability of the alternative hypothesis}

\item{Y:}{Data list of K (number of dose groups) matrices having dimension n (cells) times p (genes)}
}
\value{
A list containing the log-likelihood under the null hypothesis,
log-likelihood under the alternative hypothesis, log Bayes factor and the Bayes Factor values
}
\description{
Title Compute Bayes factors for all the p-genes
}
