% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_concordance_AUC.R
\name{calculate_concordance_AUC}
\alias{calculate_concordance_AUC}
\title{Calculate concordance AUC Refer to: Soneson, Charlotte, and Mark D. Robinson.
"Bias, robustness and scalability in single-cell differential
expression analysis." Nature methods 15, no. 4 (2018): 255.}
\usage{
calculate_concordance_AUC(DETestoutput)
}
\value{
Normed area under the concordance curve.
}
\description{
Calculate concordance AUC Refer to: Soneson, Charlotte, and Mark D. Robinson.
"Bias, robustness and scalability in single-cell differential
expression analysis." Nature methods 15, no. 4 (2018): 255.
}
\author{
Satabdi Saha
}
