% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_PR_curve.R
\name{compute_PR_curve}
\alias{compute_PR_curve}
\title{Title Computes PR curve for DEG classification}
\usage{
compute_PR_curve(sim, DETestoutput)
}
\value{
PR plot
}
\description{
Title Computes PR curve for DEG classification
}
\author{
Satabdi Saha
}
