% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ROC_curves.R
\name{compute_ROC_curve}
\alias{compute_ROC_curve}
\title{Title Computes ROC curve for DEG classification}
\usage{
compute_ROC_curve(sim, DETestoutput)
}
\value{
ROC plot
}
\description{
Title Computes ROC curve for DEG classification
}
\author{
Satabdi Saha
}
