% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_group_summary.R
\name{data_group_summary}
\alias{data_group_summary}
\title{Summarises data by groups}
\usage{
data_group_summary(data)
}
\arguments{
\item{data}{a data object}
}
\value{
INSETRT DESCRIPTION
}
\description{
Summarises data by groups
}
