% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{getDEGs}
\alias{getDEGs}
\title{General function to run the statistical test abstracting much of the individual steps}
\usage{
getDEGs(
  sce,
  DETestoutput,
  threshold = 0.05,
  bayes.threshold = 1/3,
  fc.threshold = 0,
  pct.expressed = 0,
  verbose = TRUE
)
}
\arguments{
\item{sce}{SingleCellExperiment object with a logcounts assay
and Dose column in the cell metadata}

\item{method}{the statistical test(s) to run on the sce object. Can be any of the following: Bayes, LRT.linear, LRT.multiple, ANOVA, KW, WRS, MAST, ChiSQ}
}
\value{
a vector of p values from the ANOVA test
}
\description{
General function to run the statistical test abstracting much of the individual steps
}
