% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retired_functions.R
\name{KW_test}
\alias{KW_test}
\title{Performed genewise Kruskal Wallis test on a cell x row matrix}
\usage{
KW_test(data)
}
\arguments{
\item{data}{matrix where cells are rows and genes are columns}
}
\value{
a vector of p values from the Kruskal Wallis test
}
\description{
Performed genewise Kruskal Wallis test on a cell x row matrix
}
\author{
Satabdi Saha
}
