% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.lklh.marginal.alt.R
\name{log.lklh.marginal.alt}
\alias{log.lklh.marginal.alt}
\title{Title Compute the marginal log-likehihood of the p-genes under the alternative hypotheesis}
\usage{
\method{log}{lklh.marginal.alt}(par, Y)
}
\arguments{
\item{par}{Parameters of the function}
}
\value{
Marginal log-likelihood under the alternative hypothesis
}
\description{
Title Compute the marginal log-likehihood of the p-genes under the alternative hypotheesis
}
\author{
Satabdi Saha
}
