% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes_test.R
\name{log.lklh.marginal.null}
\alias{log.lklh.marginal.null}
\title{Title Compute the marginal log-likehihood of the p-genes under the null hypotheesis}
\usage{
\method{log}{lklh.marginal.null}(par, Y)
}
\arguments{
\item{par}{Parameters of the function}
}
\value{
Marginal log-likelihood under the null hypothesis
}
\description{
Title Compute the marginal log-likehihood of the p-genes under the null hypotheesis
}
\author{
Satabdi Saha
}
