% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_test.R
\name{logProd}
\alias{logProd}
\title{INSERT DESCRIPTION HERE}
\usage{
logProd(prod, logand)
}
\arguments{
\item{prod}{INSERT DESCR}

\item{logand}{INSERT DESCR}
}
\description{
INSERT DESCRIPTION HERE
}
\author{

}
