% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRT_linearModel.R
\name{LRT_linearModel}
\alias{LRT_linearModel}
\title{Title Reworked version of LRT_linearModel.R (no-loops)}
\usage{
LRT_linearModel(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object with a logcounts assay
and Dose column in the cell metadata}
}
\value{
A dataframe having the test statistic and p-values
}
\description{
Title Reworked version of LRT_linearModel.R (no-loops)
}
\author{
Satabdi
}
