% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRT_linearModel_reworked.R
\name{LRT_linearModel_reworked}
\alias{LRT_linearModel_reworked}
\title{Title Reworked version of LRT_linearModel.R (no-loops)}
\usage{
LRT_linearModel_reworked(sce)
}
\arguments{
\item{sce}{}
}
\value{
A dataframe having the test statistic and p-values
}
\description{
Title Reworked version of LRT_linearModel.R (no-loops)
}
\author{
Satabdi
}
