% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_test.R
\name{LRT_multiple_groups}
\alias{LRT_multiple_groups}
\title{INSERT DESCRIPTION HERE}
\usage{
LRT_multiple_groups(data, data_ind)
}
\arguments{
\item{data}{the list of vectors of the positive observations( must of length sum(each vector in data_ind))}

\item{data_ind}{a list having K vectors of the zeroes/ones for expressed or not in each of the K -groups}
}
\value{
INSERT RETURN DESCR HERE
}
\description{
INSERT DESCRIPTION HERE
}
\author{

}
