% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_LRT_multipleGroups.R
\name{LRT_multipleModel}
\alias{LRT_multipleModel}
\title{INSERT DESCRIPTION}
\usage{
LRT_multipleModel(data.list)
}
\arguments{
\item{data.list}{INSERT DESCRIPTION}
}
\value{
INSERT DESCRIPTION
}
\description{
INSERT DESCRIPTION
}
