% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRT_multipleModel.R
\name{LRT_multipleModelNew}
\alias{LRT_multipleModelNew}
\title{INSERT DESCRIPTION}
\usage{
LRT_multipleModelNew(data.list)
}
\arguments{
\item{data.list}{INSERT DESCRIPTION}
}
\value{
INSERT DESCRIPTION
}
\description{
INSERT DESCRIPTION
}
