% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRT_multipleModel.R
\name{lrt_norm}
\alias{lrt_norm}
\title{INSERT DESCRIPTION HERE}
\usage{
lrt_norm(data)
}
\arguments{
\item{data}{the list of vectors of the observations}

\item{data_ind}{a list having K vectors of the zeroes/ones for expressed or not in each of the K -groups}
}
\value{
INSERT RETURN DESCR HERE
}
\description{
INSERT DESCRIPTION HERE
}
\author{
Satabdi Saha
}
