% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrtest.R
\name{lrtest}
\alias{lrtest}
\title{INSERT DESCRIPTION HERE}
\usage{
lrtest(w.x, w.y, x, y)
}
\arguments{
\item{w.x}{vector of zeros/ones for expressed or not in each group}

\item{w.y}{vector of zeros/ones for expressed or not in each group}

\item{x}{vector of the positive observations (must be of length sum(w.x) and sum(w.y))}

\item{y}{vector of the positive observations (must be of length sum(w.x) and sum(w.y))}
}
\value{
INSERT RETURN DESCR HERE
}
\description{
INSERT DESCRIPTION HERE
}
\author{

}
