% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newBayesTest.R
\name{new_Bayes_factor_multiple}
\alias{new_Bayes_factor_multiple}
\title{Title Bayesian Test for dose-dependent differential gene expression analysis}
\usage{
new_Bayes_factor_multiple(Y, prior, detailed = FALSE)
}
\arguments{
\item{Y}{list of size K (nlevels(dose_level)) of n (cells) times p(gene, in our case p=1)
matrices of observed data having n cells and p genes in K treatment dose groups}

\item{prior}{INSERT DESCRIPTION}

\item{detailed}{A boolean denoting how much info to show in result}
}
\value{
output: A list having the different parts of the log-likelihood function, log
prior odds and the log Bayes factor test statistic
}
\description{
Title Bayesian Test for dose-dependent differential gene expression analysis
}
\author{
Satabdi Saha
}
