% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newBayesTest.R
\name{new_bayesDETest}
\alias{new_bayesDETest}
\title{INSERT DESCRIPTION}
\usage{
new_bayesDETest(priors, detailed = FALSE)
}
\arguments{
\item{priors}{INSERT DESCRIPTION}

\item{detailed}{A boolean denoting how much info to show in result}
}
\value{
INSESRT DESCRIPTION - describe omega etc...
}
\description{
INSERT DESCRIPTION
}
