% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retired_functions.R
\name{old_bayesDETest}
\alias{old_bayesDETest}
\title{Performs a genewise ANOVA test on a SingleCellExperiment object}
\usage{
old_bayesDETest(data.list, m, tau_k_mu, tau_mu, prior_Alter, prior_Null)
}
\arguments{
\item{sce}{SingleCellExperiment object with a logcounts assay
and Dose column in the cell metadata}
}
\value{
a vector of p values from the ANOVA test
}
\description{
Performs a genewise ANOVA test on a SingleCellExperiment object
}
