% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_test.R
\name{runAnova}
\alias{runAnova}
\title{Performs a ANOVA test on a logcounts vector for a given dose}
\usage{
runAnova(data, dose)
}
\arguments{
\item{data}{The logcounts vector}

\item{dose}{The dose to analyze}
}
\value{
A p value from the ANOVA test
}
\description{
Performs a ANOVA test on a logcounts vector for a given dose
}
