% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retired_functions.R
\name{runANOVA_KW}
\alias{runANOVA_KW}
\title{Performs a genewise ANOVA test on a SingleCellExperiment object}
\usage{
runANOVA_KW(logcounts, cell.meta)
}
\arguments{
\item{sce}{SingleCellExperiment object with a logcounts assay
and Dose column in the cell metadata}
}
\value{
a vector of p values from the ANOVA test
}
\description{
Performs a genewise ANOVA test on a SingleCellExperiment object
}
