% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KW_test.R
\name{runKW}
\alias{runKW}
\title{Performs a Kruskal Wallis test on a logcounts vactor for a given dose}
\usage{
runKW(data, dose)
}
\arguments{
\item{data}{The logcounts vector}

\item{dose}{The dose to analyze}
}
\value{
A p value from the Kruskal Wallis test
}
\description{
Performs a Kruskal Wallis test on a logcounts vactor for a given dose
}
\author{
Jack Dodson
}
