% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statUtils.R
\name{runMAST}
\alias{runMAST}
\title{INSERT DESCRIPTION}
\usage{
runMAST(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object with a logcounts assay
and Dose column in the cell metadata}
}
\value{
INSERT DESCRIPTION
}
\description{
INSERT DESCRIPTION
}
