% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratBimod.R
\name{runSeuratBimod}
\alias{runSeuratBimod}
\title{Performs a genewise Wilcoxon Rank Sum test on a SingleCellExperiment object}
\usage{
runSeuratBimod(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object with a logcounts assay
and Dose column in the cell metadata}
}
\value{
a vector of p values from the Wilcoxon Rank Sum test
}
\description{
Performs a genewise Wilcoxon Rank Sum test on a SingleCellExperiment object
}
