% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Chi-square-test-of-indepence-geneExp-DoseLevel.R
\name{split_tibble}
\alias{split_tibble}
\title{Chi-square test of independence for binned gene expression data}
\usage{
split_tibble(tibble, column = "col")
}
\arguments{
\item{sce}{A Single Cell Object}
}
\value{
FDR adjusted p-values
}
\description{
Chi-square test of independence for binned gene expression data
}
\author{
Satabdi Saha
}
