% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesFactorMultiple.R
\name{Bayes_factor_multiple}
\alias{Bayes_factor_multiple}
\title{Title Bayesian Test for dose-dependent differential gene expression analysis}
\usage{
Bayes_factor_multiple(
  Y,
  m,
  m_0,
  tau_k_mu = rep(1, 9),
  K,
  tau_mu = c(0.5, 1, 2, 3),
  b_sigma = 1,
  a_sigma = 6,
  a_w = 0.8,
  b_w = 0.2,
  prior_alt,
  prior_null
)
}
\arguments{
\item{Y}{: list of size K (nlevels(dose_level)) of n (cells) times p(gene, in our case p=1)
matrices of observed data having n cells and p genes in K treatment dose groups}

\item{m}{: vector of group means}

\item{m_0}{: grand_mean of m}

\item{tau_k_mu}{}

\item{K}{: number of dose groups}

\item{tau_mu}{}

\item{b_sigma}{}

\item{a_sigma}{}

\item{a_w}{}

\item{b_w}{}

\item{prior_alt}{}

\item{prior_null}{}
}
\value{
output: A list having the different parts of the log-likelihood function, log
prior odds and the log Bayes factor test statistic
}
\description{
Title Bayesian Test for dose-dependent differential gene expression analysis
}
\author{
Satabdi Saha
}
