% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prior_hyperparameters.R
\name{calc_a_sigma_b_sigma}
\alias{calc_a_sigma_b_sigma}
\title{Calculate a_sigma, b_sigma}
\usage{
calc_a_sigma_b_sigma(sigma_mean, sigma_var)
}
\arguments{
\item{sigma_mean}{INSERT DESCR HERE}

\item{sigma_var}{INSERT DESCR HERE}
}
\value{
a_sigma and b_sigma values
}
\description{
Calculate a_sigma, b_sigma
}
\author{
Satabdi Saha
}
