% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes_test.R
\name{DoseMatrix2List}
\alias{DoseMatrix2List}
\title{INSERT DESCRIPTION}
\usage{
DoseMatrix2List(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object with a logcounts assay
and Dose column in the cell metadata}
}
\value{
INSERT DESCRIPTION
}
\description{
INSERT DESCRIPTION
}
