% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newBayesTest.R
\name{new_bayesDETest}
\alias{new_bayesDETest}
\title{TODO: NEEDS WORK/TESTING
Performs a genewise ANOVA test on a SingleCellExperiment object}
\usage{
new_bayesDETest(prior)
}
\arguments{
\item{sce}{SingleCellExperiment object with a logcounts assay
and Dose column in the cell metadata}
}
\value{
a vector of p values from the ANOVA test - describe omega etc...
}
\description{
TODO: NEEDS WORK/TESTING
Performs a genewise ANOVA test on a SingleCellExperiment object
}
