% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newBayesTest.R
\name{new_sceCalcPriors}
\alias{new_sceCalcPriors}
\title{INSERT DESCRIPTION}
\usage{
new_sceCalcPriors(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object with a logcounts assay
and Dose column in the cell metadata}
}
\value{
INSERT DESCRIPTION
}
\description{
INSERT DESCRIPTION
}
