% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KW_test.R, R/retired_functions.R
\name{runKW}
\alias{runKW}
\title{Performs a Kruskal Wallis test on a logcounts vactor for a given dose}
\usage{
runKW(logcounts, cell.meta)

runKW(logcounts, cell.meta)
}
\arguments{
\item{data}{The logcounts vector}

\item{dose}{The dose to analyze}

\item{sce}{SingleCellExperiment object with a logcounts assay
and Dose column in the cell metadata}
}
\value{
A p value from the Kruskal Wallis test

a vector of p values from the ANOVA test
}
\description{
Performs a Kruskal Wallis test on a logcounts vactor for a given dose

Performs a genewise ANOVA test on a SingleCellExperiment object
}
\author{
Jack Dodson
}
