% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRS_test.R
\name{runWRS}
\alias{runWRS}
\title{Performs a Wilcoxon Rank Sum test on a logcounts vactor for a given dose}
\usage{
runWRS(data, dose)
}
\arguments{
\item{data}{The logcounts vector}

\item{dose}{The dose to analyze}
}
\value{
A p value from the Wilcoxon Rank Sum test
}
\description{
Performs a Wilcoxon Rank Sum test on a logcounts vactor for a given dose
}
