% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes_test.R, R/retired_functions.R
\name{sceCalcPriors}
\alias{sceCalcPriors}
\title{INSERT DESCRIPTION}
\usage{
sceCalcPriors(sce)

sceCalcPriors(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object with a logcounts assay
and Dose column in the cell metadata}
}
\value{
INSERT DESCRIPTION

a vector of p values from the ANOVA test
}
\description{
INSERT DESCRIPTION

Performs a genewise ANOVA test on a SingleCellExperiment object
}
