% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes_test.R
\name{sceCalcPriors}
\alias{sceCalcPriors}
\title{INSERT DESCRIPTION}
\usage{
sceCalcPriors(sce, fixed.priors = TRUE)
}
\arguments{
\item{sce}{SingleCellExperiment object with a logcounts assay
and Dose column in the cell metadata}
}
\value{
INSERT DESCRIPTION
}
\description{
INSERT DESCRIPTION
}
