% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retired_functions.R
\name{WRS_test}
\alias{WRS_test}
\title{Performed genewise Wilcoxon Rank Sum test on a cell x row matrix}
\usage{
WRS_test(data, dose_level_ref, dose_level_test)
}
\arguments{
\item{data}{matrix where cells are rows and genes are columns}
}
\value{

}
\description{
Performed genewise Wilcoxon Rank Sum test on a cell x row matrix
}
